internationalization: 🌐 add i18n definitions and dictionaries
Add new files and functions to support internationalization (i18n) in the application. This includes the addition of the Locales enum in the i18n.definitions.ts file, the getDictionary function in the app/dictionaries.ts file, and the useGetDictionary hook in the hooks/useDictionaries.ts file. The dictionaries for the English and Spanish languages have also been added in the app/dictionaries/en.json and app/dictionaries/es.json files respectively.